Splitting up Storybook stories into separate files (#858)
* Splitting up Storybook stories into separate files

* Disabling TODO checks for CodeClimate

* Removing published build

* Fixing lint errors

* Fixing comment

* Fixing karma test; fixing peer dependency requirement

* Adding expects to capybara tests

* Adding helper for tests to check page transitions

* Fixing merge conflicting dangling line

* Updating scripts, fixing component names

* Fixing defaultProps for Dropdown

* Fixing missing photo and webpack config

* Trigger codeclimate

* Adding missing EOL

* Fixing storybook webpack config

* Updating loofah gem for circleci
