+++ .gitignore
@@ -12,7 +12,7 @@
 /db/*.sqlite3-journal
 
 # Ignore all logfiles and tempfiles.
-/log/*.log
+*.log
 /tmp
 
 *.DS_Store
