+++ client/.storybook/webpack.config.js
@@ -1,6 +1,17 @@
+const path = require('path');
+
 const baseConfig = require('../webpack.config.base');
 
 module.exports = Object.assign(baseConfig, {
+
+  resolve: {
+    alias: baseConfig.resolve.alias,
+    extensions: baseConfig.resolve.extensions,
+    modules: [
+      'client/app',
+      'node_modules',
+    ],
+  },
 
   module: {
     rules: [
@@ -9,18 +20,18 @@ module.exports = Object.assign(baseConfig, {
         loader: 'style-loader!css-loader?modules&camelCase&importLoaders=1&localIdentName=[name]__[local]___[hash:base64:5]!sass-loader',
       },
       {
-        test:/\.ya?ml$/,
-        loader: 'yml-loader'
+        test: /\.ya?ml$/,
+        loader: 'yml-loader',
       },
       {
-        test: /\.(png|jp(e*)g|svg)$/,  
+        test: /\.(png|jp(e*)g|svg)$/,
         use: [{
-            loader: 'url-loader',
-            options: { 
-                limit: 8000,
-                name: 'images/[hash]-[name].[ext]'
-            } 
-        }]
+          loader: 'url-loader',
+          options: {
+            limit: 8000,
+            name: 'images/[hash]-[name].[ext]',
+          },
+        }],
       },
     ],
   },
