+++ client/README.md
@@ -1,4 +1,6 @@
-Client folder generated by the React on Rails gem.
+# React on Rails
+
+This `client` folder is generated by the React on Rails gem.
 
 See documentation [at github.com/shakacode/react_on_rails](https://github.com/shakacode/react_on_rails) for details on how it is organized.
 
@@ -7,3 +9,17 @@ If you need additional help, please consider:
 * [Our ShakaCode Forum for React on Rails](https://forum.shakacode.com/c/rails/reactonrails).
 * Joining our Slack discussion room by [emailing us a bit about you and your project](mailto:contact@shakacode.com).
 * [Hiring us](https://forum.shakacode.com/c/rails/reactonrails) for coaching and custom web application development for your project.
+
+# Storybook
+
+You can develop React components inside Storybook. To set up Storybook, you will need to set up the
+_if me_ application first (see the root [README.md](../README.md) for installation instructions).
+
+Then from the project root directory, execute the following:
+
+```
+rake i18n:js:export
+bin/start_app
+# stop the process via CTRL+C
+cd client && yarn && yarn storybook
+```
