+++ client/app/bundles/shared/components/Dropdown/DropdownFillSmall.jsx
@@ -2,4 +2,5 @@ import Dropdown from './Dropdown';
 import css from './Dropdown.scss';
 
 const DropdownFillSmall = Dropdown(css.fill);
+DropdownFillSmall.displayName = 'DropdownFillSmall';
 export default DropdownFillSmall;
