+++ client/app/bundles/shared/components/Dropdown/DropdownGhost.jsx
@@ -2,5 +2,6 @@ import Dropdown from './Dropdown';
 import css from './Dropdown.scss';
 
 const DropdownGhost = Dropdown(css.ghost);
+DropdownGhost.displayName = 'DropdownGhost';
 export default DropdownGhost;
