+++ client/app/bundles/shared/components/Dropdown/DropdownGhostSmall.jsx
@@ -2,5 +2,5 @@ import Dropdown from './Dropdown';
 import css from './Dropdown.scss';
 
 const DropdownGhostSmall = Dropdown(css.ghost_small);
+DropdownGhostSmall.displayName = 'DropdownGhostSmall';
 export default DropdownGhostSmall;
-
