+++ client/app/bundles/shared/components/Dropdown/index.jsx
@@ -1,0 +1,4 @@
+export { default as Dropdown } from './Dropdown';
+export { default as DropdownFillSmall } from './DropdownFillSmall';
+export { default as DropdownGhost } from './DropdownGhost';
+export { default as DropdownGhostSmall } from './DropdownGhostSmall';
