+++ client/package.json
@@ -75,6 +75,7 @@
     "flow-bin": "^0.54.1",
     "flow-typed": "^2.3.0",
     "jasmine": "^2.7.0",
+    "jasmine-core": "^3.1.0",
     "jasmine-enzyme": "^3.7.0",
     "karma": "^1.7.0",
     "karma-jasmine": "^1.1.0",
