+++ client/webpack.config.base.js
@@ -3,9 +3,9 @@ const path = require('path');
 module.exports = {
   resolve: {
     alias: {
-      libs: path.resolve(path.join(process.cwd(), 'app', 'libs')),
-      config: path.resolve(path.join(process.cwd(), '..', 'config')),
       app: path.resolve(path.join(process.cwd(), '..', 'app')),
+      config: path.resolve(path.join(process.cwd(), '..', 'config')),
+      libs: path.resolve(path.join(process.cwd(), 'app', 'libs')),
     },
     extensions: ['.js', '.jsx'],
   },
