+++ package.json
@@ -3,7 +3,8 @@
   "private": true,
   "dependencies": {},
   "scripts": {
-    "postinstall": "cd client && yarn install"
+    "postinstall": "cd client && yarn install",
+    "start": "foreman start -f Procfile.dev"
   },
   "engines": {
     "node": "6.x"
